Add define to allow inferring BRAM for memory by not resetting it.
With the requirement to reset memory on reset, yosys can't use iCE40
BRAM. This is challenging to route for higher number of LEDs. By
avoiding the requirement it will now fail formal verification, but
will place and route much faster and more efficiently.

Adding top that changes the values demonstrates an example that permit
elimination of the memory altogether.